requirements.txt
Flask>=2.0.0
google-auth-oauthlib>=1.2.0  # Latest version for OAuth flow
google-auth>=2.35.0           # Latest version for credentials
google-api-python-client>=2.150.0  # Latest version (removes file_cache warnings)
pyOpenSSL>=22.0.0
google-genai>=0.2.0
python-dotenv>=1.0.0
python-decouple>=3.8
Flask-CORS>=4.0.0
Flask-Talisman>=1.1.0
redis>=4.5.0
rq>=1.15.0
SQLAlchemy>=2.0.0
Flask-SQLAlchemy>=3.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
Flask-Caching>=2.0.0
tenacity>=8.2.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
prometheus-client>=0.19.0
structlog>=23.2.0
python-json-logger>=2.0.7

